MonteCarloPropFirmSimulator/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
pydantic
python-multipart
openpyxl
matplotlib
yfinance
psycopg2-binary
